fix: split up core tests
